refactor: split catalog response decoding
